requirements.txt
fastapi
uvicorn
deepface
Pillow
